Move generate_genesis_config to eth2.configs
