+++ eth2/configs.py
@@ -1,9 +1,24 @@
 from dataclasses import Field, dataclass, fields
-from typing import Collection, Dict, Iterable, Tuple, Union, cast
+from typing import Any, Collection, Dict, Iterable, Tuple, Union, cast
 
+from eth.constants import ZERO_HASH32
 from eth_utils import decode_hex, encode_hex, to_dict
+from ssz.tools.dump import to_formatted_dict
+from typing_extensions import Literal
 
-from eth2.beacon.typing import Epoch, Gwei, Second, Slot
+from eth2.beacon.genesis import initialize_beacon_state_from_eth1
+from eth2.beacon.state_machines.forks.serenity.configs import SERENITY_CONFIG
+from eth2.beacon.state_machines.forks.skeleton_lake.config import (
+    MINIMAL_SERENITY_CONFIG,
+)
+from eth2.beacon.tools.builder.initializer import (
+    create_genesis_deposits_from,
+    create_key_pairs_for,
+    mk_genesis_key_map,
+    mk_withdrawal_credentials_from,
+)
+from eth2.beacon.tools.misc.ssz_vector import override_lengths
+from eth2.beacon.typing import Epoch, Gwei, Second, Slot, Timestamp
 
 ConfigTypes = Union[Gwei, Slot, Epoch, Second, bytes, int]
 EncodedConfigTypes = Union[str, int]
@@ -98,3 +113,45 @@ class Eth2Config:
     def from_formatted_dict(cls, data: Dict[str, EncodedConfigTypes]) -> "Eth2Config":
         # NOTE: mypy does not recognize the kwarg unpacking here...
         return cls(**_decoder(data, fields(cls)))  # type: ignore
+
+
+def generate_genesis_config(
+    config_profile: Literal["minimal", "mainnet"],
+    genesis_time: Timestamp = None,
+) -> Dict[str, Any]:
+    eth2_config = _get_eth2_config(config_profile)
+    override_lengths(eth2_config)
+    validator_count = eth2_config.MIN_GENESIS_ACTIVE_VALIDATOR_COUNT
+
+    validator_key_pairs = create_key_pairs_for(validator_count)
+    deposits = create_genesis_deposits_from(
+        validator_key_pairs,
+        withdrawal_credentials_provider=mk_withdrawal_credentials_from(
+            eth2_config.BLS_WITHDRAWAL_PREFIX.to_bytes(1, byteorder="little")
+        ),
+        amount_provider=lambda _public_key: eth2_config.MAX_EFFECTIVE_BALANCE,
+    )
+    eth1_block_hash = ZERO_HASH32
+    eth1_timestamp = eth2_config.MIN_GENESIS_TIME
+    initial_state = initialize_beacon_state_from_eth1(
+        eth1_block_hash=eth1_block_hash,
+        eth1_timestamp=Timestamp(eth1_timestamp),
+        deposits=deposits,
+        config=eth2_config,
+    )
+
+    if genesis_time:
+        initial_state.set("genesis_time", genesis_time)
+    # instead of genesis_state = adjust_
+
+    return {
+        "eth2_config": eth2_config.to_formatted_dict(),
+        "genesis_validator_key_pairs": mk_genesis_key_map(
+            validator_key_pairs, initial_state
+        ),
+        "genesis_state": to_formatted_dict(initial_state),
+    }
+
+
+def _get_eth2_config(profile: str) -> Eth2Config:
+    return {"minimal": MINIMAL_SERENITY_CONFIG, "mainnet": SERENITY_CONFIG}[profile]
