+++ trinity/main_validator.py
@@ -10,13 +10,12 @@ from eth2.beacon.tools.builder.initializer import load_genesis_key_map
 from eth2.beacon.tools.misc.ssz_vector import override_lengths
 from eth2.beacon.types.states import BeaconState
 from eth2.beacon.typing import Slot
-from eth2.configs import Eth2Config
+from eth2.configs import Eth2Config, generate_genesis_config
 from eth2.validator_client.cli_parser import parse_cli_args
 from eth2.validator_client.config import Config
 from eth2.validator_client.tools.directory import create_dir_if_missing
 from trinity._utils.logging import LOG_FORMATTER
 from trinity.bootstrap import load_trinity_config_from_parser_args
-from trinity.components.eth2.network_generator.component import _generate_genesis_config
 from trinity.config import ValidatorClientAppConfig
 from trinity.constants import APP_IDENTIFIER_VALIDATOR_CLIENT
 
@@ -61,7 +60,7 @@ def main_validator() -> None:
             validator_client_app_config.genesis_config_path
         )
     except FileNotFoundError:
-        genesis_config = _generate_genesis_config("minimal")
+        genesis_config = generate_genesis_config("minimal")
 
     eth2_config = Eth2Config.from_formatted_dict(genesis_config["eth2_config"])
     override_lengths(eth2_config)
